ref(onboarding): Split chalice onboarding docs
